backend/agent-core/pyproject.toml
[project]
name = "agent-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.5",
    "pydantic>=2.12.5",
    "selenium>=4.40.0",
    "sentence-transformers>=5.2.2",
    "uvicorn>=0.40.0",
    "webdriver-manager>=4.0.2",
    "beautifulsoup4>=4.12.0",
    "python-multipart>=0.0.22",
    "pypdf>=6.7.0",
    "qdrant-client>=1.16.2",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that need browser/network (run with -m integration)",
]
testpaths = ["tests"]
